:warning: :sparkles: Add support to crop the printing
This commit adds the capability to crop the printing to fit a
specific screen size. Two keywords were added to control this
functionality: `screen_size` to pass the screen size and `crop` to
define the crop behavior.

This is a breaking change because, by default, the data will be
cropped.

Closes #11