:bug: Add sortkeys to _pretty_table
The keyword `sortkeys` is used only when printing `Dict` to sort
the keys. However, if a global configuration of this keyword is set
using `@ptconf`, then every `@pt` will fail if the element is not a
`Dict`. This was caused because the function `_pretty_table` does
not have that keyword. Hence, it was added in this commit. Notice
that this keyword does not have any function inside
`_pretty_table`.

A test for this function was also added.