:sparkles: Give priority to Tables.jl API when printing
The Tables.jl API is not the priority when checking if a variable
can be printed as a tables.

This fixes the problem related to `StructArray`. This type is an
`AbstractVecOrMat` and also complies with Tables.jl API. However,
the old version gave preference to the `AbstractVecOrMat` type,
leading to a strange printing.

Closes #28