:warning: :wrench: Modify the behavior of hlines in Text backend
The keyword `hlines` in Text back-end now controls all the
horizontal lines, including the top line, header line, and the
bottom line. Thus, the variables `top_line`, `header_line`, and
`bottom_line` of the structure `TextFormat` were removed. This
structure now has a variable `hlines` that defines the common
behavior. This behavior will be used if the user pass the option
`hlines = nothing` to `pretty_table`.

This is a breaking change.