:warning: :sparkles: Add default option of `vlines` to TextFormat
TextFormat can now have a default selection of `vlines` just as in
`hlines`. With this commit, the sole case in which `left_border`
and `right_border` is needed can be replicated by using
`vlines = [:begin,:end]`. Thus, those options were removed from
`TextFormat`.

This can be a breaking change if the user defined a custom format.