:wrench: Improve cropping mechanism in text back-end
The vertical cropping mechanism now takes into account the lines inside
the cells. Hence, we can process fewer lines is cropping is used, saving
computational workload.

The algorithm was also improved to check if only one line is left before
cropping. In this case, the last line is drawn instead of the
continuation line.