:wrench: Do not pre-store the length of table data
In Text backend, we were creating two matrices `data_len` and
`header_len` to store the cell lengths. However, for very big matrices,
this requires a lot of allocations, leading to a bad performance. The
downside is that we need to compute the text width of the cells many
times, but it is very fast. In the end, this new algorithm allocates
less and it is faster.