:warning: Remove filters
Filters were a feature very few people use and caused a lot of problems
to add new features. Since we can replicate the result using standard
Julia functions (like `filter` and `map`), we decided to remove it from
PrettyTables.jl to make it easier to maintain the code. For more
information, please, refer to:

    https://discourse.julialang.org/t/question-for-prettytables-jl-users-can-we-remove-filters/82609/8