:wrench: Make easier to build column labels
By default, we now have `merge_column_label_cells = :auto`. Hence, the
user can pass the objects to merge the column cells without any
additional option.

The algorithm was improve to add very little overhead if the column
label vector does not have any elements of type `MultiColumn` or
`EmptyCells` on it.

Furthermore, if the user pass a vector of elements instead of a vector of
a vector, the algorithm treat as a column label with only one line.